shared freelist
Now that freelist is an FFI struct, this commit adds the concept of "named freelists".
If the call to `freelist.new()` includes a name, the module saves the freelist together
with the struct type in a private list.  If the name was already known, the same freelist
is returned.

The `share_lists()` and `receive_shared()` work together at the outside and inside of
a state sandbox to share the list of named freelists.

The packet freelist creation from `core.packet` is moved from the main module chunk to
a new `packet.init()` function, that should be called before allocating packets, but
after the named freelists had been shared.

When creating a state sandbox, after loading the standard libraries and the main chunk,
the app_loader calls `packet.init()` (to be sure the packet freelist is created) and then
uses `freelist.share_lists()` to inject the freelists to the `freelist.receive_shared()`
inside the sandbox.  Then `packet.init()` is called inside, so the packet module ends up
receiving the very same freelist as was created at the outside.