register.lua: Remove jit.off() voodoo
This was added long ago when counter values were "sticking"
suspiciously. We have no way to reproduce this problem at present and
since it does not appear in recent testing we remove this kludge.

The worst anticipated regression would be that counter registers stop
updating. If that happens then we will need to revisit the issue.