git-workflow.md: Edited and expanded
Spell out the activities that contribute to Snabb Switch maintenance.

Revise the description of how to submit a change. Emphasise that the
goal is to land the change on a branch that somebody is going to merge
with master and once that is done you can mostly relax. (You don't
have to push your change for a whole month until it lands in a release
- only until somebody else merges it onto a branch that they are going
to push upstream.)

Rephase from role-oriented ("Casual users should do this") to
task-oriented ("How to download the latest release"). Ideally
everybody will be a user, developer, and maintainer to varying
extents.