core.shm: Add alias() and children()
alias(toname, fromname) creates a new name for an existing
object. This is implemented as a symlink.

children() lists the names of objects under a path. This is a
directory listing.

The initial use for these features is providing links to apps. The app
has to learn what links it has using children() and each link it finds
is an alias of one created by the engine.

For example, the link 'a.tx->b.rx' would be one shm object created by
the engine with a symlink to make it available to each app:

    /engine/link/a.tx->b.rx  (real struct link created by engine)
    /app/a/link/output/tx -> /engine/link/a.tx->b.rx
    /app/b/link/input/rx  -> /engine/link/a.tx->b.rx