packetblaster: Fix to avoid starvation
This commit introduces two changes to avoid the NIC becoming idle:

- Use maximum descriptor ring size for 82599 (32K descriptors)
- Busywait instead of sleeping between transmissions

This makes packetblaster now run at 100% CPU as seen in top.

Testing on Interlaken this boosted transmission from 3 Mpps to 12.8
Mpps. (I am not sure why it is less than 14.8 Mpps. Could be that we
need to poke a TX register differently to get maximum packet rate.)