intel1g: Draft of Intel 1G (I350, etc) driver
This is a work in progress published only for information sharing.

The main intention is to support the Intel 1G network card family. I
am also using this module to explore some ideas around alternative
coding style and app configuration parameters. These ideas may or may
not make it into the final version.