vhost.h, vhost_user.h: Use more explicit types
Switch to stricter C declarations to ensure that the memory layout of
vhost-user requests matches the vhost-user specification. For example,
use 'uint32_t' instead of 'unsigned int'. Add __attribute__((packed))
to struct declarations.

This is to ensure that the on-the-wire memory layout follows the
vhost-user spec rather than the C compiler.

vhost-user spec:
https://github.com/qemu/qemu/blob/master/docs/specs/vhost-user.txt