core.link: Streamline counters
Reduce the amount of counter-related code and data in `struct link`.

The ring indexes (read, write) and now upgraded to 64-bit integers that
do not wrap around [*]. This means that we don't need to use separate
packet tx/rx counters because the ring indexes serve that purpose
directly. (We do have to be careful to handle wrap-around when actually
looking up elements in the ring based on the index.)

The byte counters are simply removed. Do we really need them?

Dropped packets are still counted directly because this is important
information and I don't immediately see a way to derive it indirectly.

[*] 64-bit packet counters really should not wrap around. Even at one
billion packets per second it would take more than 500 years. However,
if this is not a safe assumption then we could detect wrap around and
account for it on the counters.

(Google: "2^64/(1 billion per second" => "584.554531 years".)