Speed up Murmur hash functions
Performance of lib.hash.murmur is improved by reducing the number of
memory accesses to process the "tail" portion of the data.  Current
benchmarks indicate that a pass over a 6-byte value (e.g. an Ethernet
MAC address) takes about 15 CPU cycles with MurmurHash3_x64_128 and
about 19 cycles with MurmurHash3_x86_32, provided that all constants
are pre-loaded into registers.

The base class is changed to make sure that the same CTYPE is used for
allocating the storage that holds the result of a hash function.  The
previous code, which generated distinct CTYPEs, lead to subtle
performance issues due to side-exits being taken in compiled code when
CTYPE-checking guarded instructions failed.