src/Makefile: Add 'make submods'
This target updates and rebuilds the required git submodules. The
Makefile tells you to run this when it detects that the submodule
versions are not in sync with the source code.

This is a more streamlined version of the previous behavior where the
Makefile would suggest running a series of commands to update your
submodules. Now 'make submods' has the same effect.