core.memory: Catch SIGSEGV to map DMA memory [PoC]
This is a quick and dirty proof-of-concept hack done to support
SnabbCo/snabbswitch#565 (multicore tests):

Extend core.memory so that the virtual addresses of all packets are
valid in all Snabb Switch processes. The memory allocated by
memory.dma_alloc() is treated specially: if another process tries to
access it then this triggers a SIGSEGV handler that will detect which
memory is needed and automatically map it into the process.

Two simple test scripts are provided: s1 that allocates memory and s2
that modifies it without any explicit mapping.

Run like this:

    $ sudo ./snabb snsh s1
    Installing SIGSEGV handler
    address cdata<char *>: 0x5009f7c00000
    current value = 0
    current value = 0
    current value = 0
    ...

Then use the same address to change the value with s2 that simply
converts it to a pointer and pokes it:

    $ make && sudo strace  ./snabb snsh s2 0x5009f7c00000 42

and now you see the new value in s1:

    current value = 42

This is really quick and dirty! I will clean it up if it works and we
decide to keep it.