snabbmark: Add first step towards porting basic1 to asm
Super early sketch of the code! shared just for fun.

The idea is simple: clone the 'basic1' benchmark in assembler to see
what performance is possible. Then we could use this as a benchmark
for deciding how optimized our Lua code is. Could be that in writing
the assembler code we get some bright ideas for improving the Lua code
too.

This is very rough code that implements:

    source -> sink

and recycles packets via a freelist.

The full implementation would have to implement the real basic1
topology i.e. include a "Tee" app that duplicates packets.

This is not optimized, etc, so don't draw any sweeping conclusions yet :-).

Note also that I did most of my assembler programming 20+ years ago on
the MC68000 so it will take a little practice to find a decent style :).