Avoid ctype diversity for registers in intel10g
The application of lib.bounds_checked() to the cdata objects used for
receive and transmit descriptor rings causes branches in JIT traces
for different instances of the driver, leading to highly sub-optimal
side-traces in applications that use multiple instances in the same
program.  The patch only suggest a minimal solution by disabling the
bounds checking logic entirely.