snabbmark_asm_lua: Port of asm code to Lua
Port the asm version of 'snabbmark basic1' back to Lua.

The assembler version was simplified in various ways e.g. it does not
have a real breathe loop and it assumes that links/freelists never
overflow or underflow.

This is a fairly faithful port of that simplified assembler version
back to Lua. The purpose is to see apples-to-apples how fast the Lua
code runs when you 'cheat' in the same ways as the assembler code.

Loosely the results seem to be (on Interlaken):

    original:  27 Mpps
    asm:      140 Mpps
    lua-asm:   62 Mpps