pmu: Minor improvements to reporting interface
Renamed 'ref-cycles' to 'ref_cycles' both for consistency with other
counters and to make it easier to use as a Lua table key.

report() now takes a table for input rather than a counterset. This
makes it easy to write Lua code that manipulates data (e.g. taking
deltas from separate runs) and then call report() to format it.

report() now lexically sorts the counters based on their names, with
the exception of the fixed-purpose counters (cycles, instructions,
ref_cycles) that are printed first in a fixed order. This is intended
to increase consistency and make it slightly easier to compare results
by eyeball.