appbench: New subcommand of snabbmark program
appbench measures the impact on CPU performance counters when a new
app is introduced to an app network between a Source and a Sink.

This can be used to analyze the behavior of an individual app: for
example how many cycles per packet it consumes, how many accesses to
each level of the cache heirarchy per packet, and how many branch
mispredictions per packet.

Here is the example analysis of the Tee app processing 100M packets:

    EVENT                                             TOTAL     /packet
    cycles                                    1,409,813,298      14.098
    ref_cycles                                1,056,440,016      10.564
    instructions                              3,858,706,478      38.587
    br_misp_retired.all_branches                    361,693       0.004
    mem_load_uops_retired.l1_hit              1,324,505,106      13.245
    mem_load_uops_retired.l2_hit                109,539,516       1.095
    mem_load_uops_retired.l3_hit                    180,222       0.002
    mem_load_uops_retired.l3_miss                         0       0.000
    packet                                      100,000,000       1.000