snsh: Added '-q' to break program to REPL via keyboard
Now you can run any Snabb-based program in a mode that supports
interrupting it with the keyboard (Control-\ i.e. SIGQUIT) such that
it leaves the breathe loop and enters an interactive read-eval-print
loop (REPL). Here you can poke around with arbitrary Lua expressions
and finally return back to the engine by pressing Control-d (EOF).

The argument -q (--sigquit-repl) enables a timer callback in the
engine that checks if SIGQUIT has been received and starts a
REPL. This can be combined with the -p (--program) argument to run
another program such as packetblaster or snabbnfv.

For example:

  snabb snsh -q -p packetblaster replay 64.cap 0000:01:00.0

will run packetblaster but with the hook enabled such that each time
you press Control-\ the process will receive a SIGQUIT and enter the
REPL like this:

    0000:01:00.1    TXDGPC (TX packets)     13,154,784      GOTCL (TX octets)       894,522,388
    ^\[snsh: SIGQUIT caught - entering REPL]
    Snabb>

Note: SIGQUIT is detected by a hook function running on a timer and so
it is only effective when the engine's breathe loop is running. (The
REPL does not run inside a Unix signal handler -- the signal only sets
a flag saying that it should be started.)