lib/pmu_pebs: PEBS (Precise Event Based Samping) [sketch]
This does not work yet but it is an implementation of the CPU "PEBS"
feature. I am not sure whether it is easy or hard to complete this:
the hardware is simple but can this really be used from Linux userspace?

More details from the comments in pmu_pebs.lua:

    PEBS is a hardware feature where the CPU collects its own samples
    and writes them into memory. The samples collected by the CPU can
    include interesting details such as the data address that is being
    referenced when a cache miss occurs. Events can be triggered based
    on a subset of performance events (cache misses, etc).

    This module sets up PEBS and enables event logging into memory.

    Refer to Intel's system programming manual (vol.3) for details:
    http://www.intel.com/content/www/us/en/processors/architectures-software-developer-manuals.html