core.engine: Enable selftest() method
The selftest method has been suppressed in a previous change. It seems
better to run it and for all available features (including those not
enabled by default).

Currently fails in the app restart-on-error part of the tests. This
feature is not currently enabled by default but it should still be
working and tested.