Handle reassembly of IPv4 fragmented packets in the lwAFTR
This uses the "fragmentv4" module to collect packets fragments received
from the IPv4 side of the network, and reassemble them. The fragment cache
works in the same way as the IPv6 version: a table in the application
state keep a list of fragments for each (src_ip, dst_ip, frag_id) combination.
On each fragment, reassembly is attempted, and cached fragments cleared either
when reassembly is successful, or when an invalid set of fragments is
detected.

Also, this patchs adds a "fragmentv4.is_ipv4_fragment()" helper function,
which is used to determine whether a packet should go into the fragment cache.