Fix bugs in PMU code
There were two bugs in the PMU code.  The first was that if the events
you select matched more than pmu_x86_ngeneral counters, the set of
events was being incorrectly truncated.  This would later manifest as an
error when we try to do defs[set[j+1]], as some random element in the
middle of `set' could be nil.

The other bug was that the global IA32_PERF_GLOBAL_CTRL register wasn't
being initialized.  One symptom could be that any given counter wasn't
counting -- for me it was the cycles fixed-function counter.  We didn't
observe this in the past because this is global state that could have
gotten initialized by any other program -- perhaps a previous invocation
of `perf`.