VirtioNet: introduce the user-space virtio-net driver
An implementation of the virtio-net driver specification, intended to run
inside a guest virtual machine.

Using indirect descriptors allow for more effective usage of the descriptor
table. Each statically allocated to a indirect descriptor table, where a
preallocated packet header is placed. Adding a buffer is simply finding the
index of the descriptor and putting the buffer physical address into its addr field.