Use new parser for lwaftr config
This CL adapts the lwaftr to rely on the new parser.  As such some of
the configuration is simplified -- i.e. no need to call strange library
functions, and no need to do any ad-hoc validation of the range of vlan
tags.  At the same time, we switch to specify the binding table by name
in the config instead of by value.  This will allow for a switch to the
new address map and binding table format, and removes the need for --bt
options to various scripts.  All documentation, tests, and benchmarks
have been updated.