Use LuaJIT htons/htonl
LuaJIT would have to save everything around a call to C.htons.  Instead
if we use bitops and the ffi, we can inline it.

Note that there is a htonl implementation in core.lib, but it can return
negative integers, which is not what we want.

Also use ffi.fill when we can instead of memset, for the same reason.