nDPI FFI binding usage example using libpcap
The "examples/readpcap" script implements an example on how to use the "ndpi"
module to implement a tool similar to the "pcapReader" one included with nDPI
in Lua. The main difference is that only reading packets from a pcap capture
is implemented (live capture is not supported), and the statistics are printed
differently.

As a freebie, the "examples.pcap" module is provided, which implements a very
minimal wrapper over libpcap which can be used to iterate over the packets of
a pcap file.