Globally use '#!/usr/bin/env <shell>' hashbang line
Convert all shell scripts from having path dependencies like:
  #!/bin/bash
to
  #!/usr/bin/env bash

The immediate utility of this change is to fix annoying breakage on
NixOS. For example, now "make doc/snabbswitch.html" works.

Exception is that I did not descend into subdirectories of
dependencies e.g. ljsyscall and pflua.

I made the changes with an Emacs macro in the hope of avoiding an
unfortunate typo that will break something.