src/README.md: Simplify core.memory API
Streamline the core.memory API:

- Rewrite the blurb.
- Keep dma_alloc() and virtual_to_physical().
- Remove dma_min_addr and dma_max_addr (unsed).
- Deprecate huge_page_size (used but should it be?)

Backstory here is that the core.memory module has exported low-level
information to other modules: the VFIO module (since removed) that
needed to whitelist all DMA memory with the kernel, a /dev/vhost-net
accelerator for tap devices (since removed) that also needed to
whitelist all packet memory, and to a lesser extent the vhost-user
implementation (that is still alive and well but may be overusing the
memory API e.g. in dead code that should be removed).

Adding the deprecation note on "huge_page_size" may or may not be a good
idea. Have to look more closely at the usage and clarify with a later
commit.