Pass only interface name to syscall.util.if_nametoindex()
The syscall.util.if_nametoindex() only needs to be passed the name of the
interface. The version of the function which accepts an open socket as second
argument is internal to ljsyscall, and not even exported to client code.

Also, this moves the call to syscall.util.if_nametoindex() to the top of
RawSocket:new() to simplify error handling: this way one less explicit call
to syscall.close() is needed in case of errors.