core.memory: Rewrite memory.c in Lua using ljsyscall
Switch to allocating HugeTLB pages as mmaped files on hugetlbfs
instead of POSIX shared memory. The advantage of POSIX shared memory
was that we did not need to rely on having a hugetlbfs filesystem
mounted -- but now we auto-mount one at /var/run/snabb/hugetlbfs.

This design is a bit simpler and should also make it easier for one
Snabb process to map a page that was allocated by a different process.

The main motivation for rewriting memory.c is to make it easier to
maintain. Lua code is more in keeping with the rest of Snabb and it
doesn't require people to be expert C programmers and debate the
merits of POSIX compliance and using GCC extensions and so on. (It
also puts a bit more distance between Snabb and monstrosities like
autoconf which cannot be a bad thing.)