scripts/lock: Script for locking lab servers
Introduce a 'lock' command for synchronizing tests on lab servers.

Users would use 'lock' when they are running a command that is likely
to conflict with other users e.g. a snabb process that accesses PCI
devices or a long running test.

The script can be used for a single command:

  lock sudo ./snabb ...
  lock sudo ./mytestscript

Or to start a shell that holds the lock (1h idle timeout):

  lock
  $

This script would need to be used in a consistent and disciplined way
by all users of a given machine to be effective.