Clear out Flags in ICMPv4 echo-reply
It's a common practice in ping packets to set the Flags field of the
IPv4 header with value DF (Don't Fragment). The reason for that is
implementation of Path MTU discovery [1]

Echo-reply packets are generated by copying an echo-request payload
and swapping IP addresses plus changing message type. The Flags field
should be cleared out in the echo-reply packet.

[1] https://en.wikipedia.org/wiki/Path_MTU_Discovery