src/Makefile: Flatten Lua module names (no directory)
This change puts all Lua module names at the top-level and ignores
their directory prefixes. So instead of `core.link` we have `link` and
instead of `program.snabbnfv.traffic.traffic` we have `traffic`.

This is based on a suggestion by @capr on #734.

The basic idea here is to make the directory structure a pure
implementation detail that is not visible to users. This is motivated
by the idea that users don't care about the internal structure and
implementors prefer more flexibility about how code is organized.

One way this could simplify the organization of code would be to simply have:

    src/lwaftr

instead of

    src/program/lwaftr
    src/apps/lwaftr

i.e. an obvious place to collect all of the code that is specific to
one application independent of the content of that code (apps,
libraries, command-line programs, etc).

One way it would simplify the API is by hiding the distinction between
`core` and `lib` that is perhaps more of an implementation detail anyway.

To make this change actually work we would need to:

- Give every module a unique name.
- Update `require()` lines.
- Invent new rules for finding programs (and perhaps apps).

Some thought would be needed to decide how to do that and how this
change relates to the broader topic of #734.