Documentation: Remove build artifacts from repo
The markdown files in the repository are now all "source" e.g.
containing ascii art rather than references to Ditaa-generated images.

Specifically this change overwrites the ".md" files with their
".src.md" versions and removes the ".png" image files.

This stuff can all be built on demand instead of stored in the repo.