doc/template.latex: Updated to suit recent pandoc
The LaTeX template in our tree did not work with recent versions of
pandoc. The problem is essentially this one:
http://tex.stackexchange.com/questions/257418/error-tightlist-converting-md-file-into-pdf-using-pandoc

This change manually replaces the generic parts of the template with
the reference version found here:
https://github.com/jgm/pandoc-templates/blob/master/default.latex

This may make the build incompatible with older versions of pandoc, I
am not sure. I did consider dropping the custom LaTeX template
entirely but decided on balance that it looks nicer and is worth
spending a little effort to try and maintain.