scripts/process-markdown: New takes output file arg
Previously process-markdown took only an input file argument and
printed the result to standard output. Now it takes both an input file
and output file argument.

The important change is that the '.images' directory containing ditaa
images is now created in the same directory as the output file, and
that directory is created if necessary.

Previously the '.images' was created in the source directory and in
practice this was only the right thing if the output file happened to
also be in that directory. This was true the past when we were
converting foo.src.md -> foo.md but is not true anymore now that we
convert foo.md -> obj/foo.md.

src/Makefile is updated to reflect this. This also made it possible to
drop the Makefile dependency from the manual on the 'snabb'
executable, which was an ugly hack intended to ensure that the
required directory tree in obj/ had been created somehow. Now the
process-markdown script automatically creates any required output
directory.