Combine intermediate hash result while iterating over the input
This makes the hash functions generated by make_cdata_hash_function()
use the intermediate value calculated for the previous 32-bit chunks
of the input combined with the output for the current chunk. Without
this fix, Changes at the beginning of the input would not make the
final hash result vary, which was incorrect behavior.