Add on-a-stick mode
The on-a-stick mode allows to run the lwAFTR in one single NIC.  Incoming
traffic is categorized per packet type (v4 or v6) and forwarded to
the corresponding link. The lwAFTR runs normally on each incoming link.
Outgoing packets are joined together and redirected to the NIC so they
can be transmitted. Example:

sudo ./snabb lwaftr run -v --conf lwaftr.conf --on-a-stick 83:00.0

In addition, the on-a-stick mode features a mirroring option:

sudo ./snabb lwaftr run -v --conf lwaftr.conf --on-a-stick 83:00.0
--mirror tap0

Packets which IPv4 source and destination address match
/var/run/snabb/PID/v4v6_mirror are copied to the 'mirror' output link,
which is linked to an TAP interface set by --mirror.