PodHashMap: Be more like CTable
Eventually we need to replace PodHashMap with CTable, and part of that
will be porting things from PodHashMap *to* ctable (streaming lookup and
the ability to load and save).  But there are some things that CTable
does that we don't do, like iteration, and to port that feature here we
make a refactor to allocate all memory using a "calloc" helper that will
one day be refactored elsewhere into some nice library.