connectx4: Progress on initialization routine
Commands are being successfully executed towards the card.

The full initialization procedure is not in place yet. Support for
commands that span multiple input/output pages needs to be implemented.

Current expected behavior when running the selftest is to successfully
execute the commands ENABLE_HCA, QUERY_ISSI, QUERY_PAGES, MANAGE_PAGES,
and then to fail in QUERY_HCA_CAP (likely because it has multipage
output).