core.timer: Inhibit debug messages
Print debug messages only when enabled by editing the source file, not
when _G.developer_debug is enabled.

The debug mode prints a message every time a timer is executed. This
could easily produce thousands of messages per second in reasonable
applications. This seems appropriate when specifically debugging the
timer module but not when enabling a more general debug mode.