Expose the nDPI version and use versioned ndpi.protocol_ids
This obtains at runtime the version of the nDPI library using the
ndpi_revision() function, and exposes it as a table at "ndpi.lib_version".
This table has fields .major, .minor, and .patch with their usual meanings.

Additionally, the existing ndpi.protocol_ids module is renamed to
ndpi.protocol_ids_1_7, and a new ndpi.protocol_ids_1_8 module is added
(created using tools/update-protocol-ids). The detected library version
numbers are used to determine which one of the versioned protocol_ids
submodule has to be loaded.

This patch is a first step towards supporting nDPI 1.8 (issue #1).